requirements.txt
Django==3.1.13
django-improved-user==1.0.0
django-widget-tweaks==1.4.5

openpyxl==3.0.3
python-dateutil==2.8.1
